+++ core/initializer/test_suite_metadata_acquirer/test_suite_metadata_acquirer.go
@@ -32,6 +32,8 @@ Spins up a testsuite container in test-listing mode and returns the "set" of tes
 func GetAllTestNamesInSuite(
 		testSuiteImage string,
 		dockerManager *commons.DockerManager) (map[string]bool, error) {
+	parentContext := context.Background()
+
 	// Create the tempfile that the testsuite image will write test names to
 	testNamesFp, err := ioutil.TempFile("", "test-names")
 	if err != nil {
@@ -47,7 +49,7 @@ func GetAllTestNamesInSuite(
 	containerLogFp.Close()
 	defer os.Remove(containerLogFp.Name())
 
-	bridgeNetworkIds, err := dockerManager.GetNetworkIdsByName(context.Background(), bridgeNetworkName)
+	bridgeNetworkIds, err := dockerManager.GetNetworkIdsByName(parentContext, bridgeNetworkName)
 	if err != nil {
 		return nil, stacktrace.Propagate(
 			err,
@@ -63,7 +65,7 @@ func GetAllTestNamesInSuite(
 	bridgeNetworkId := bridgeNetworkIds[0]
 
 	testListingContainerId, err := dockerManager.CreateAndStartContainer(
-		context.Background(),
+		parentContext,
 		testSuiteImage,
 		// TODO parameterize these
 		bridgeNetworkId,
@@ -86,7 +88,7 @@ func GetAllTestNamesInSuite(
 	}
 
 	testListingExitCode, err := dockerManager.WaitForExit(
-		context.Background(),
+		parentContext,
 		testListingContainerId)
 	if err != nil {
 		banner_printer.PrintContainerLogsWithBanners(logrus.StandardLogger(), testListingContainerDescription, containerLogFp.Name())